Move tests into proper folder structure Make Async tests for create and update The maxSizeInMegabytes is known to fail, and is being pushed to the feature branch for investigation (due to 16*actual-size partition interaction)
